requirements.txt
google-api-python-client
google-auth
nltk
transformers
torch
sentencepiece
tqdm
pyyaml
click
numpy
matplotlib
pandas
